fix: 인플루언서 등록 RegisterInfluencerDto에 profile_image추가, @RequestPart와 @RequestPart로 나누어져있던 파라미터들을 @ModelAttrubute로 통합

+++ src/main/java/com/shotty/shotty/S3ImageService.java
@@ -33,7 +33,7 @@ public class S3ImageService {
 
     public String upload(MultipartFile image) {
         if(image.isEmpty() || Objects.isNull(image.getOriginalFilename())){
-            return null;
+            throw new S3Exception("비어있는 이미지 파일");
         }
         return this.uploadImage(image);
     }
@@ -97,6 +97,9 @@ public class S3ImageService {
     }
 
     public void deleteImageFromS3(String imageAddress){
+        if (imageAddress == null) {
+            return;
+        }
         String key = getKeyFromImageAddress(imageAddress);
         try{
             amazonS3.deleteObject(new DeleteObjectRequest(bucketName, key));
